remove ignis.settings; move CACHE_DIR to __init__.py
