+++ ignis/__init__.py
@@ -1,4 +1,6 @@
+import os
 import sys
+
 
 __version__ = "0.1"
 
@@ -12,3 +14,7 @@ if "sphinx" not in sys.modules:
     gi.require_version("Gdk", "4.0")
     gi.require_version("Gtk4LayerShell", "1.0")
     gi.require_version("GdkPixbuf", "2.0")
+
+    from gi.repository import GLib
+    CACHE_DIR = f"{GLib.get_user_cache_dir()}/ignis/"
+    os.makedirs(CACHE_DIR, exist_ok=True)
